+++ lib/helpers.ts
@@ -1,5 +1,5 @@
 import { AnySelector } from "./selectors";
-import { ListType, AnyControl } from "./controls";
+import { ListOf, AnyControl } from "./controls";
 
 export interface Query {
     [prop: string]: Query | AnySelector | AnyControl;
@@ -13,12 +13,12 @@ export interface SelectorResultType {
     exist: boolean;
 }
 
-export type TypeOfQuery<Q extends AnySelector | ListType<any> | Query> = {
+export type QueryOf<Q extends AnySelector | ListOf<any> | Query> = {
     0: Q extends AnySelector ? SelectorResultType[Q["type"]] : never;
-    1: Q extends ListType<infer T> ? TypeOfQuery<T>[] : never;
+    1: Q extends ListOf<infer T> ? QueryOf<T>[] : never;
     2: Q extends Query ? {
-        [prop in keyof Q]: TypeOfQuery<Q[prop]>
+        [prop in keyof Q]: QueryOf<Q[prop]>
     }: never;
 }[Q extends AnySelector ? 0
-    : Q extends ListType<any> ? 1
+    : Q extends ListOf<any> ? 1
     : Q extends Query ? 2 : boolean];